FIX: Move doc components back to frontend to fix regressions
