+++ packages/frontend/src/components/index.ts
@@ -1,3 +1,8 @@
+export * from "./completions";
 export * from "./id_input";
+export * from "./inline_input";
+export * from "./input_options";
 export * from "./json_import";
+export * from "./name_input";
+export * from "./resizable";
 export * from "./rich_text_editor";
