+++ packages/frontend/src/components/resizable.css
