+++ packages/frontend/src/components/resizable.tsx
