+++ packages/frontend/src/diagram/object_cell_editor.tsx
@@ -1,6 +1,6 @@
 import { createSignal } from "solid-js";
 
-import { NameInput } from "catcolab-ui-components";
+import { NameInput } from "../components";
 import { ObInput } from "../model/object_input";
 import type { CellActions } from "../notebook";
 import type { Theory } from "../theory";
