+++ packages/frontend/src/model/morphism_cell_editor.tsx
@@ -1,7 +1,7 @@
 import { createMemo, createSignal, useContext } from "solid-js";
 import invariant from "tiny-invariant";
 
-import { NameInput } from "catcolab-ui-components";
+import { NameInput } from "../components";
 import type { CellActions } from "../notebook";
 import type { Theory } from "../theory";
 import { LiveModelContext } from "./context";
