+++ packages/frontend/src/model/object_list_editor.tsx
@@ -11,9 +11,8 @@ import {
 } from "solid-js";
 import invariant from "tiny-invariant";
 
-import type { InputOptions } from "catcolab-ui-components";
 import type { Ob, QualifiedName } from "catlog-wasm";
-import { ObIdInput } from "../components";
+import { type InputOptions, ObIdInput } from "../components";
 import { deepCopyJSON } from "../util/deepcopy";
 import { LiveModelContext } from "./context";
 import type { ObInputProps } from "./object_input";
