+++ packages/frontend/src/page/document_page.tsx
@@ -14,11 +14,12 @@ import {
 } from "solid-js";
 import invariant from "tiny-invariant";
 
-import { IconButton, InlineInput, ResizableHandle } from "catcolab-ui-components";
+import { IconButton } from "catcolab-ui-components";
 import { type LiveAnalysisDocument, getLiveAnalysis } from "../analysis";
 import { AnalysisNotebookEditor } from "../analysis/analysis_editor";
 import { AnalysisInfo } from "../analysis/analysis_info";
 import { type Api, type DocumentType, useApi } from "../api";
+import { InlineInput, ResizableHandle } from "../components";
 import { type LiveDiagramDocument, getLiveDiagram } from "../diagram";
 import { DiagramNotebookEditor } from "../diagram/diagram_editor";
 import { DiagramInfo } from "../diagram/diagram_info";
