+++ packages/frontend/src/theory/theory.ts
@@ -1,8 +1,8 @@
 import type { DblModel, DblTheory, MorType, ObOp, ObType } from "catlog-wasm";
 
-import type { KbdKey } from "catcolab-ui-components";
 import type { DiagramAnalysisComponent, ModelAnalysisComponent } from "../analysis";
 import { uniqueIndexArray } from "../util/indexing";
+import type { KbdKey } from "../util/keyboard";
 import type { ArrowStyle } from "../visualization";
 import { MorTypeMap, ObTypeMap } from "./types";
