+++ packages/ui-components/src/index.ts
@@ -1,15 +1,9 @@
 export * from "./alert";
 export * from "./button";
-export * from "./completions";
 export * from "./dialog";
 export * from "./fixed_table_editor";
 export * from "./foldable";
 export * from "./form";
 export * from "./icon_button";
-export * from "./inline_input";
-export * from "./input_options";
-export * from "./name_input";
 export * from "./panel";
-export * from "./resizable";
 export * from "./spinner";
-export * from "./util/keyboard";
